Add files via upload
Add in three new if statements. The first one opens up my fraternity's chapter website and the second one opens the national website for the fraternity. The last if statement added was to search on google what the person ask if it does not meet any of the other if statements.